python-apt/gdebi based pbuilder-satisfydepends-gdebi (closes: #453388)
From: Michael Vogt <mvo@ubuntu.com>

In Ubuntu, we've applied the attached patch to achieve the following:

    * Makefile:
      - Addded pbuilder-satisfydepends-gdebi
    * pbuilder-buildpackage-funcs:
      - always pass --chroot to the pbuilder-satisfydepends (needed for
        pbuilder-satisfydepends-gdebi)
    * pbuilder-satisfydepends-gdebi:
      - added alternative b-d resolver

pbuilder-satisfydepends-gdebi is a build-dependencies resolver based
on gdebi (that in turn uses python-apt/libapt). It is fast and does
not need to build a dummy package. It should behave exactly like
apt-get build-dep $pkg

It used to be more interessting because of the speed gain when
pbuilder-satisfydepends-classic was the default. You probably want to
add a recommends or suggests for gdebi too (if you like the patch
method).

Modified by dancerj: --chroot is given before --chrootexec so that
cowdancer can continue to work.