satisfydepends-gdebi: reorganize code to go in-line with other funcs.
The original implementation did not work well with cowdancer, change
it so that gdebi is called inside chroot and inside chroot only.

This has the drawback that gdebi must be available within chroot, but
it is much simpler than adding conditional cowdancer support to gdebi, and hopefully better for future development.