Add support for sources.list.d and validate autodetection.
A the system might use sources.list.d, but might miss sources.list
entirely. All the relevant files are concatenated and the first deb
source is picked from that list.

When the autodection fails, make sure the responsability to insert
the valid information drops on the shoulders of the user.

Even if the shell is non-interactive, add some mirror.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>