Dereference symlinks when copying hooks. (closes: #499358)
I keep a number of different hooks in my $HOME/.pbuilder/hooks/ 
directory, and some are symlinks to files outside that directory.

When 'pbuilder-runhooks' copies the hooks, it copies them without 
dereference, which fails inside the chroot since the referenced files 
are not found.

The attached patch, generated using 'git format-patch', fixes this by 
using the '-L' ('--dereference') option to 'cp' when copying the hook 
files into the pbuilder chroot.