Allow preseeding of pbuilder/mirrorsite.  Closes: #598316
As far as I can tell, the db_set calls all through pbuilder.config were
mostly to get around the default value -- but there's no need for it, given
that it will be set correctly if we're non-interactive anyway.

I've tested this patch in all the different ways I can think of, to exercise
all the code paths, and it works in the way I'd expect (a sensible-looking,
valid mirror is set).  Absent a formal test plan, I can't be any more
confident that it does the right thing in all situations.