Ensure that nothing is still mounted before deleting the build chroot.  Closes: #542915
Whilst pbuilder is quite careful to minimise the risk of accidental data
removal (with find -xdev, unmounting everything it mounts, etc), there is
one case that it doesn't handle -- directories on the same filesystem as the
chroot being bind-mounted into the chroot, because -xdev doesn't recognise
the mount as a separate filesystem and pbuilder doesn't know to unmount it
because it didn't mount it in the first place.

This patch acts as a final safety net, to avoid attempting to delete
anything if there is anything at all still mounted in the chroot.