Fix pbuilder-satisfydepends-classic for arch selectors.  Ref: #363193
This was reported as a side-effect of #363193
(http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=363193#191), but the
problem was actually caused by p-sd-classic passing in a mangled version of
the dependency info (spaces were replaced by forward slashes), which the old
checkbuilddeps_archdeps handled because it didn't parse the package spec so
much as grep it into submission.  All good now.