Check correct architecture for checkarchitecture, and make it more testable
The checkarchitecture function was examining and reporting against the main
host architecture, instead of the chroot architecture.  Whilst this isn't
usually a problem, an i386 chroot on an amd64 machine might give odd
results.

I also took the liberty of moving the arch detection out of the function, so
that we can more easily test it by just passing in an arch of our choosing,
rather than having to rely on the kindness of strangers.