Switch all dpkg-architecture calls to include -f
Would you believe that the *default* behaviour of dpkg-architecture is to
pick up existing environment variables and use them in favour of what's
either the "natural" value of the setting, or what you've EXPLICITLY SET ON
THE COMMAND LINE (with, for instance, the -a option).

Whilst this is undeniably insane, it's apparently intended behaviour, so
there's not a lot we can do about it other than passing -f every time we
call dpkg-architecture.  RAAAAAAAGE.